+++ agent/controller/agent.go
@@ -22,9 +22,9 @@ import (
 	"github.com/hoophq/hoop/agent/config"
 	"github.com/hoophq/hoop/agent/controller/awseks"
 	"github.com/hoophq/hoop/agent/controller/featureflagstate"
+	"github.com/hoophq/hoop/agent/controller/system/bareexec"
 	"github.com/hoophq/hoop/agent/controller/system/dbprovisioner"
 	"github.com/hoophq/hoop/agent/controller/system/pgmanager"
-	"github.com/hoophq/hoop/agent/controller/system/resourcemanager"
 	"github.com/hoophq/hoop/agent/controller/system/runbookhook"
 	"github.com/hoophq/hoop/agent/rds"
 	"github.com/hoophq/hoop/agent/secretsmanager"
@@ -256,11 +256,8 @@ func (a *Agent) processPacket(pkt *pb.Packet) {
 	case pbsystem.RunbookHookRequestType:
 		runbookhook.ProcessRequest(a.client, pkt)
 
-	case pbsystem.ResourceManagerRequestType:
-		resourcemanager.ProcessRequest(a.client, pkt)
-
 	case pbsystem.BareExecRequestType:
-		resourcemanager.ProcessRequest(a.client, pkt)
+		bareexec.ProcessRequest(a.client, pkt)
 
 	case pbsystem.PgManagerPlanRequestType:
 		pgmanager.ProcessPlanRequest(a.client, pkt)
